refactor more actions from components
