+++ src/components/empty-slot/inventory-slot.png
