+++ src/features/app-state/actions/opt-out.js
@@ -1,0 +1,9 @@
+
+export default function optOut() {
+  return dispatch => {
+    dispatch({
+      type: 'OPT_OUT_DOWNLOAD',
+      payload: {}
+    });
+  }
+}
