+++ src/features/dialog-manager/dialogs/inventory-dialog/index.js
@@ -10,36 +10,29 @@ import './styles.scss';
 
 const InventoryDialog = () => {
 
-  const [viewItem, setViewItem] = useState(null);
-
-  function handleViewItem(item) {
-    setViewItem(
-      <ViewItem
-        data={item}
-        onClose={handleCloseItem} />
-    );
-  }
-
-  function handleCloseItem() {
-    setViewItem(null);
-  }
+  const [viewItem, setViewItem] = useState(false);
 
   return(
     <Dialog>
 
-      { viewItem }
+      <ViewItem
+        open={Boolean(viewItem)}
+        data={viewItem}
+        onClose={() => setViewItem(false)} />
 
       <div className='flex-row inventory-dialog__container'>
 
-        <div className='flex-column inventory-dialog__child'
-          style={{width: '40%'}}>
+        <div className='flex-column inventory-dialog__child' style={{width: '40%'}}>
           <EquippedItems />
         </div>
 
-        <div className='flex-column inventory-dialog__child'
-          style={{width: '60%'}}>
-          <div className='inventory-dialog__backpack' style={{backgroundImage: `url(${Backpack})`}}>
-            <BackpackItems view_item={handleViewItem} />
+        <div className='flex-column inventory-dialog__child' style={{width: '60%'}}>
+          <div className='inventory-dialog__backpack'
+            style={{backgroundImage: `url(${Backpack})`}}>
+
+            <BackpackItems
+              view_item={(item) => setViewItem(item)} />
+
           </div>
         </div>
