+++ src/features/inventory/actions/equip-item.js
